+++ jscomp/test/pipe_send_readline.ml
@@ -1,18 +1,17 @@
-
 (*  should give a warning on unused attribute..   [@@mel.xx] *)
 
 
 type readline
 external on :
   ([ `line of (string -> unit [@u])
-   | `close of (unit -> unit [@u])]
+   | `close of (unit -> unit [@u0])]
      [@mel.string]) ->
   'self =
   "on" [@@mel.send.pipe:readline as 'self]
 let u rl =
   rl
   |> on (`line (fun [@u] x -> Js.log x ))
-  |> on (`close (fun [@u] () -> Js.log "finished"))
+  |> on (`close (fun [@u0] () -> Js.log "finished"))