+++ jscomp/test/tramp_fib.ml
@@ -5,7 +5,7 @@ let test_id = ref 0
 let eq loc x y = Mt.eq_suites ~test_id ~suites loc x y
 
 
-type 'a bounce = Continue of 'a | Suspend of (unit -> 'a bounce [@u])
+type 'a bounce = Continue of 'a | Suspend of (unit -> 'a bounce [@u0])
 (* https://eli.thegreenplace.net/2017/on-recursion-continuations-and-trampolines/ *)
 (* http://gallium.inria.fr/seminaires/transparents/20141027.Frederic.Bour.pdf *)
 (* http://www.usrsb.in/blog/blog/2012/08/12/bouncing-pythons-generators-with-a-trampoline/ *)
@@ -17,7 +17,7 @@ let rec fib n k =
     (* Suspend (fun [@u]() -> k (Continue 1 ) [@u]) *)
     k 1 [@u]
   | _ ->
-    Suspend (fun [@u] () ->
+    Suspend (fun [@u0] () ->
       fib (n-1) (fun [@u] v0 ->
         fib (n-2) (fun [@u] v1 ->
           k (v0 + v1) [@u]
@@ -37,7 +37,7 @@ let u = fib 10 (fun [@u] x -> Continue x)
 let rec iter (bounce : 'a bounce) : 'a =
   match bounce with
   | Continue v -> v
-  | Suspend f -> iter (f () [@u])
+  | Suspend f -> iter (f () [@u0])
 
 
 (* first it needs to be tailcall *)
@@ -45,7 +45,7 @@ let rec isEven n =
   match n with
   | 0 -> Continue true
   | 1 -> Continue false
-  | _ -> Suspend (fun [@u] () -> isOdd (n - 1))
+  | _ -> Suspend (fun [@u0] () -> isOdd (n - 1))
 and isOdd n =
   match n with
   | 0 -> Continue false