+++ ppx/ast_core_type.mli
@@ -32,3 +32,8 @@ val make_obj : loc:Location.t -> object_field list -> core_type
 val get_uncurry_arity : core_type -> int option
 (** returns 0 when it can not tell arity from the syntax. [None] means not a
     function *)
+
+val get_uncurry_arity_from_attribute :
+  zero_arity:bool -> core_type -> int option
+(** returns 0 when it can not tell arity from the syntax. [None] means not a
+    function *)