+++ ppx/ast_uncurry_apply.mli
@@ -29,6 +29,7 @@ open Import
 val uncurry_fn_apply :
   Location.t ->
   Ast_traverse.map ->
+  zero_arity:bool ->
   expression ->
   (Asttypes.arg_label * expression) list ->
   expression_desc